requirements.txt
cloud-sql-python-connector[pymysql]
SQLAlchemy==2.0.7
python-dotenv
yfinance
pyarrow
prefect
pandas
polars